Refactor grid search & update PCC, add ipykernel
Moved the optimal bin number finder to the grid search part.
Changed PCC to take all of the columns into account as the binning will be happening at a later stage.
Implemented the grid search, the training flow, and the final testing.
Added `ipykernel` to the `requirements.txt` file to provide the `IPython` kernel.